Implement status text with scrolling

+++ component/main/os.c
@@ -7,17 +7,17 @@
 
 const char* LOG_DOMAIN = "SSD1306";
 
-void ssd1306_log_set_level(unsigned level)
+void ssd1306_log_set_level(uint8_t level)
 {
 	esp_log_level_set(LOG_DOMAIN, (esp_log_level_t)level);
 }
 
-unsigned ssd1306_log_get_level()
+uint8_t ssd1306_log_get_level()
 {
 	return esp_log_level_get_timeout(LOG_DOMAIN);
 }
 
-void ssd1306_log(unsigned level, const char* function, int line, const char* format, ...)
+void ssd1306_log(uint8_t level, const char* function, int line, const char* format, ...)
 {
 	static const char PREFIXES[] = { '\0', 'E', 'W', 'I', 'D', 'V' };
 	static const char* COLORS[] = {
